refactor(bench): reuse buildBenchmarkRunSeeds in assistant runner
Addresses cursor[bot] review on #508. Extracts the seed-sequence
generator out of `benchmark.ts` into `run-seeds.ts` (avoids the
`benchmark.ts -> registry.ts -> runner.ts -> benchmark.ts` circular
import path) and has the Assistant shared runner delegate seed
generation to it. `benchmark.ts` still re-exports the function so the
public surface is unchanged.
