+++ packages/bench/tsconfig.json
@@ -20,6 +20,7 @@
     "src/benchmark.ts",
     "src/registry.ts",
     "src/reporter.ts",
+    "src/run-seeds.ts",
     "src/scorer.ts",
     "src/ingestion-scorer.ts",
     "src/adapters/**/*.ts",
